# docs/swagger.yaml
definitions:
  gin.H:
    additionalProperties: {}
    type: object
...
  university_system_internal_domain_models.Course:
    properties:
      code:
        type: string
      created_at:
        type: string
      credits:
        type: integer
      deleted_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      teacher_id:
        type: integer
      updated_at:
        type: string
    type: object
...
  university_system_internal_domain_models.Student:
    properties:
      birthdate:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      email:
        type: string
      faculty:
        type: string
      firstname:
        type: string
      id:
        type: string
      lastname:
        type: string
      password:
        type: string
      role:
        type: string
      student_year:
        type: integer
      updated_at:
        type: string
      username:
        type: string
    type: object
  university_system_internal_domain_models.Teacher:
    properties:
      birthdate:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      department:
        type: string
      email:
        type: string
      firstname:
        type: string
      id:
        type: string
      lastname:
        type: string
      password:
        type: string
      position:
        type: string
      role:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: n_rakhat@kbtu.kz
    name: Rakhat Nurtileu
    url: http://linkedin.com/in/nurtileu-rakhat-02aba3294/
  description: System для управления пользователями, курсами и ролями
  title: University System
  version: "1.0"
paths:
...
  /courses:
    get:
      description: Возвращает список всех курсов в системе
      parameters:
      - description: Bearer токен
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/university_system_internal_domain_models.Course'
            type: array
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Получить все курсы
      tags:
      - courses
    post:
      consumes:
      - application/json
      description: Создает новый курс в системе
      parameters:
      - description: Bearer токен
        in: header
        name: Authorization
        required: true
        type: string
      - description: Данные курса
        in: body
        name: course
        required: true
        schema:
          $ref: '#/definitions/university_system_internal_domain_models.Course'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/university_system_internal_domain_models.Course'
        "400":
          description: Ошибка валидации
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Создать курс
      tags:
      - courses
  /courses/{id}:
    delete:
      description: Удаляет курс из системы по его уникальному идентификатору
      parameters:
      - description: Bearer токен
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID курса
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: Курс удален
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Неверный ID
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Удалить курс
      tags:
      - courses
    get:
      description: Возвращает курс по его уникальному идентификатору
      parameters:
      - description: Bearer токен
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID курса
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/university_system_internal_domain_models.Course'
        "400":
          description: Неверный ID
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Курс не найден
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Получить курс по ID
      tags:
      - courses
    put:
      consumes:
      - application/json
      description: Обновляет информацию о существующем курсе
      parameters:
      - description: Bearer токен
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID курса
        in: path
        name: id
        required: true
        type: integer
      - description: Новые данные курса
        in: body
        name: course
        required: true
        schema:
          $ref: '#/definitions/university_system_internal_domain_models.Course'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/university_system_internal_domain_models.Course'
        "400":
          description: Ошибка валидации
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Обновить курс
      tags:
      - courses
  /courses/{id}/students:
    get:
      description: Возвращает список студентов, записанных на данный курс
      parameters:
      - description: Bearer токен
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID курса
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/university_system_internal_domain_models.Student'
            type: array
        "400":
          description: Неверный ID
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Получить студентов курса
      tags:
      - courses
  /courses/{id}/teachers:
    get:
      description: Возвращает список преподавателей, ведущих данный курс
      parameters:
      - description: Bearer токен
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID курса
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/university_system_internal_domain_models.Teacher'
            type: array
        "400":
          description: Неверный ID
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Получить преподавателей курса
      tags:
      - courses
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"